pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["David Escobar <david.escobar@ug.uchile.cl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.9"
aiohttp = "3.8.1"
aiosignal = "1.2.0"
apscheduler = "3.6.3"
async-timeout = "4.0.2"
attrs = "22.1.0"
backoff = "2.1.2"
beautifulsoup4 = "4.11.1"
bleach = "5.0.1"
botocore = "1.27.40"
cachetools = "4.2.2"
certifi = "2022.6.15"
charset-normalizer = "2.1.0"
cycler = "0.11.0"
dataframe-image = "0.1.1"
defusedxml = "0.7.1"
entrypoints = "0.4"
fastjsonschema = "2.16.1"
fonttools = "4.34.4"
frozenlist = "1.3.0"
gql = "3.4.0"
graphql-core = "3.2.1"
idna = "3.3"
jinja2 = "3.1.2"
jmespath = "1.0.1"
jsonschema = "4.8.0"
jupyter-client = "7.3.4"
jupyter-core = "4.11.1"
jupyterlab-pygments = "0.2.2"
kiwisolver = "1.4.4"
lxml = "4.9.1"
markupsafe = "2.1.1"
matplotlib = "3.5.2"
mistune = "0.8.4"
multidict = "6.0.2"
nbclient = "0.6.6"
nbconvert = "6.5.0"
nbformat = "5.4.0"
nest-asyncio = "1.5.5"
numpy = "1.23.1"
packaging = "21.3"
pandas = "1.4.3"
pandocfilters = "1.5.0"
pillow = "9.2.0"
pygments = "2.12.0"
pyparsing = "3.0.9"
pyrsistent = "0.18.1"
python-dateutil = "2.8.2"
python-decouple = "3.6"
python-telegram-bot = "^20.3"
pytz = "2022.1"
pytz-deprecation-shim = "0.1.0.post0"
pyzmq = "23.2.0"
requests = "2.28.1"
requests-toolbelt = "0.9.1"
six = "1.16.0"
soupsieve = "2.3.2.post1"
tinycss2 = "1.1.1"
tornado = "6.2"
traitlets = "5.3.0"
tzdata = "2022.1"
tzlocal = "4.2"
urllib3 = "1.26.11"
webencodings = "0.5.1"
websockets = "10.3"
yarl = "1.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
